https://takumi01.backlog.com/view/BUS_BOOKING_SYSTEM-14 バス削除機能完成

+++ root/backend/src/main/java/vn/com/vti/bus/backend/controller/BusDeleteController.java
@@ -1,5 +1,7 @@
 package vn.com.vti.bus.backend.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,7 +9,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import vn.com.vti.bus.entity.Bus;
+import vn.com.vti.bus.entity.Route;
+import vn.com.vti.bus.entity.RouteExample;
 import vn.com.vti.bus.mapper.BusMapper;
+import vn.com.vti.bus.mapper.RouteMapper;
 
 @Controller
 @RequestMapping("/busDelete")
@@ -17,11 +22,14 @@ public class BusDeleteController {
 	
 	@Autowired BusListController buslistController;
 	
+	@Autowired
+	private RouteMapper routeMapper;
+	
 	@RequestMapping("confirm")
 	public String Confirm(@RequestParam String busId, Model model) {
 		if(busId.isEmpty()) {
-			buslistController.index(model);
-			return "";
+			return buslistController.index(model);	
+			
 			
 		}
 		Bus bus = busMapper.selectByPrimaryKey(Integer.parseInt(busId));
@@ -30,9 +38,28 @@ public class BusDeleteController {
 		
 	}
 	@RequestMapping("delete")
-	public String Delete(@RequestParam String busId) {
-		busMapper.deleteByPrimaryKey(Integer.parseInt(busId));
-		 return "redirect:/busList/index";
+	public String Delete(@RequestParam String busId, Model model) {
+		
+		RouteExample routeExample = new RouteExample();
+		routeExample.createCriteria().andBusIdEqualTo(Integer.parseInt(busId));
+		List<Route> routeList = routeMapper.selectByExample(routeExample);
+		
+		if(routeList==null || routeList.size()==0) {
+			busMapper.deleteByPrimaryKey(Integer.parseInt(busId));
+			return "redirect:/busList/index";
+			
+		}else {
+			String errMsg = "削除できません。";
+			model.addAttribute("errMsg",errMsg);
+			return buslistController.index(model);	
+		}
+	
+		
+		
+		 
+		
+		 
+		
 	}
 	
 }
